feat: add the ability for crops to have more than 1 growing phase (#473)
this change converts the `cropTimetable` property into a `cropTimeline` which allows for each crop to have a varying number of growing phases, each represented by a different image. the majority of crops have been updated with new art to take advantage of this new capability. 

this commit also has several updates to fix type warnings, and improve tests that were mocking crops and constants where it was not necessary.

* refactor(types): fix all type errors for utils/index.js

---------

Co-authored-by: Jeremy Kahn <jeremyckahn@gmail.com>