feat(closes #387): rename id to playerId
Refactor: Rename id to playerId

This commit refactors the codebase to use `playerId` instead of `id`
when referring to the player's ID.

This change was made to improve the consistency and readability of the
code. The following files were affected:
- src/components/CowCard/CowCard.js
- src/components/CowCard/Subheader/Subheader.js
- src/components/ChatRoom/ChatRoom.js
- src/game-logic/reducers/processCowBreeding.js
- src/game-logic/reducers/purchaseCow.js
- src/components/CowCard/CowCard.test.js

Refactor: Rename id to playerId

This commit refactors the codebase to use `playerId` instead of `id`
when referring to the player's ID.

This change was made to improve the consistency and readability of the
code. The following files were affected:
- src/components/CowCard/CowCard.js
- src/components/CowCard/Subheader/Subheader.js
- src/components/ChatRoom/ChatRoom.js
- src/game-logic/reducers/processCowBreeding.js
- src/game-logic/reducers/purchaseCow.js
- src/components/CowCard/CowCard.test.js
- src/components/CowCard/Subheader/Subheader.test.js
- src/components/CowPen/CowPen.js
- src/components/CowPen/CowPen.test.js

Refactor: Rename id to playerId

This commit refactors the codebase to use `playerId` instead of `id`
when referring to the player's ID.

This change was made to improve the consistency and readability of the
code. The following files were affected:
- src/components/CowCard/CowCard.js
- src/components/CowCard/Subheader/Subheader.js
- src/components/ChatRoom/ChatRoom.js
- src/game-logic/reducers/processCowBreeding.js
- src/game-logic/reducers/purchaseCow.js
- src/components/CowCard/CowCard.test.js
- src/components/CowCard/Subheader/Subheader.test.js
- src/components/CowPen/CowPen.js
- src/components/CowPen/CowPen.test.js
- src/farmhand.d.ts
- src/handlers/peer-events.js
- src/templates.js
- src/game-logic/reducers/prependPendingPeerMessage.js
- src/test-utils/stubs/peerMetadataStub.js
- src/game-logic/reducers/prependPendingPeerMessage.test.js
- src/game-logic/reducers/sellKeg.test.js

Refactor: Rename id to playerId

This commit refactors the codebase to use `playerId` instead of `id`
when referring to the player's ID.

This change was made to improve the consistency and readability of the
code. The following files were affected:
- src/components/CowCard/CowCard.js
- src/components/CowCard/Subheader/Subheader.js
- src/components/ChatRoom/ChatRoom.js
- src/game-logic/reducers/processCowBreeding.js
- src/game-logic/reducers/purchaseCow.js
- src/components/CowCard/CowCard.test.js
- src/components/CowCard/Subheader/Subheader.test.js
- src/components/CowPen/CowPen.js
- src/components/CowPen/CowPen.test.js
- src/farmhand.d.ts
- src/handlers/peer-events.js
- src/templates.js
- src/game-logic/reducers/prependPendingPeerMessage.js
- src/test-utils/stubs/peerMetadataStub.js
- src/game-logic/reducers/prependPendingPeerMessage.test.js
- src/game-logic/reducers/sellKeg.test.js
- src/components/OnlinePeersView/OnlinePeersView.js
- src/components/OnlinePeersView/OnlinePeersView.test.js
- src/components/OnlinePeersView/OnlinePeer/OnlinePeer.js
- src/components/OnlinePeersView/OnlinePeer/OnlinePeer.test.js